Add gallery app and cleanup: - Move gallery to separate app. - Adjust admin panel them logo and favicon names. - Automate process for creating user in setup.
