+++ blog/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.conf import settings
 from django_summernote.admin import SummernoteModelAdmin
 
-from .models import Post, Contact, PhotoGallery
+from .models import Post, Contact
 
 
 class PostAdmin(SummernoteModelAdmin):
@@ -10,10 +10,5 @@ class PostAdmin(SummernoteModelAdmin):
     prepopulated_fields = { 'slug': ('title',), }
 
 
-class GalleryAdmin(admin.ModelAdmin):
-    prepopulated_fields = { 'slug': ('country',), }
-
-
 admin.site.register(Post, PostAdmin)
-admin.site.register(PhotoGallery, GalleryAdmin)
 admin.site.register(Contact)
