+++ blog/fixtures/admin_interface_theme_Prokope.json
@@ -7,12 +7,12 @@
         "title": "Admin Panel",
         "title_color": "#7248C7",
         "title_visible": true,
-        "logo": "admin-interface/logo/Logo_Title.png",
+        "logo": "admin-interface/logo/logo.png",
         "logo_color": "#FFFFFF",
         "logo_max_width": 200,
         "logo_max_height": 100,
         "logo_visible": true,
-        "favicon": "admin-interface/favicon/favicon_PL6erBv.ico",
+        "favicon": "admin-interface/favicon/favicon.ico",
         "env_name": "",
         "env_color": "#E74C3C",
         "env_visible_in_header": true,
