+++ blog/models.py
@@ -4,8 +4,6 @@ from django.utils import timezone
 
 from taggit.managers import TaggableManager
 
-from photologue.models import Gallery, Photo
-
 
 # Blog Page
 class Post(models.Model):
@@ -29,22 +27,6 @@ class Post(models.Model):
 
     def __str__(self):
         return self.title
-
-
-# Gallery Page
-# TODO: Change format to get Photos directly from Photologue, then generate new
-#   to enforce pagination
-class PhotoGallery(models.Model):
-    galleries = models.ManyToManyField(Gallery)
-    country = models.CharField(max_length=199)
-    content = models.TextField()
-    slug = models.SlugField(max_length=250, allow_unicode=True, blank=True)
-
-    class Meta:
-        ordering = ["country"]
-
-    def __str__(self):
-        return f"{self.country}"
 
 
 # Contact Page
