+++ blog/urls.py
@@ -8,10 +8,8 @@ from .views import search
 
 urlpatterns = [
     path('update_server/', views.update, name='update'),
-    path('blog/', views.PostList.as_view(), name='blog'),
-    path('blog/<slug:slug>/', views.PostDetail.as_view(), name='post_detail'),
-    path('gallery/', views.CountryGallery.as_view(), name='gallery'),
-    path('gallery/<slug:slug>/', views.CountryGalleryDetail.as_view(), name='gallery_detail'),
+    path('', views.PostList.as_view(), name='blog'),
+    path('<slug:slug>/', views.PostDetail.as_view(), name='post_detail'),
     path('tags/<slug:tag_slug>/', views.TagIndexView.as_view(), name='posts_by_tag'),
     path('search/', search, name='search'),
     path('favicon.ico', RedirectView.as_view(url=staticfiles_storage.url('favicon/favicon.ico'))),
