+++ gallery/templates/gallery_detail.html
