+++ prokope/settings.py
@@ -37,8 +37,9 @@ INSTALLED_APPS = [
     'hcaptcha',
     'crispy_forms',
     'django_summernote',
-    'blog',
     'index',
+    'blog',
+    'gallery',
     'photologue',
     'sortedm2m',
     'django_otp',
